[BACKLOG-5807] (David Santos) Converted Chart classes to View classes and Axes to Base. Renamed Chart files to View author: davidmsantos90 <david.santos@webdetails.pt>
