+++ package-res/resources/web/pentaho/visual/ccc/_axes/ColumnAxis.js
@@ -14,12 +14,12 @@
 * limitations under the License.
 */
 define([
-    "./DiscreteAxis"
+  "./DiscreteAxis"
 ], function(DiscreteAxis) {
 
-    return DiscreteAxis.extend({
-        methods: {
-            defaultRole: 'columns'
-        }
-    });
+  "use strict";
+
+  return DiscreteAxis.extend({
+    defaultRole: "columns"
+  });
 });
