+++ package-res/resources/web/pentaho/visual/ccc/_axes/DiscreteAxis.js
@@ -1,87 +1,87 @@
 /*!
-* Copyright 2010 - 2015 Pentaho Corporation.  All rights reserved.
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with the License.
-* You may obtain a copy of the License at
-*
-* http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*/
+ * Copyright 2010 - 2015 Pentaho Corporation.  All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 define([
-    "cdf/lib/CCC/def",
-    "cdf/lib/CCC/pvc",
-    "./AbstractAxis",
-    "../util"
+  "cdf/lib/CCC/def",
+  "cdf/lib/CCC/pvc",
+  "./AbstractAxis",
+  "../util"
 ], function(def, pvc, AbstractAxis, util) {
 
-    return AbstractAxis.extend({
-        methods: {
-            _nonMultiGemFilter: function(gem) {
-                return gem.role !== this.chart._multiRole;
-            },
+  "use strict";
 
-            _isNullMember: function(complex, gem) {
-                var atom = complex.atoms[gem.cccDimName];
-                return util.isNullMember(atom.value);
-            },
+  return AbstractAxis.extend({
+    _nonMultiGemFilter: function(gem) {
+      return gem.role !== this.chart._multiRole;
+    },
 
-            _buildGemHtmlTooltip: function(lines, complex, context, gem, index) {
-                /*
-                 * Multi-chart formulas are not shown in the tooltip
-                 * They're on the small chart's title.
-                 *
-                 * Also, if the chart hides null members,
-                 * don't show them in the tooltip.
-                 * Using the scene's group, preferably, because the datum (here the complex) may have dimensions
-                 * that are null in the groups' own atoms.
-                 */
-                if(this._nonMultiGemFilter(gem) &&
-                   !(this.chart._hideNullMembers && this._isNullMember(context.scene.group || complex, gem))) {
-                    this.base.apply(this, arguments);
-                }
-            },
+    _isNullMember: function(complex, gem) {
+      var atom = complex.atoms[gem.cccDimName];
+      return util.isNullMember(atom.value);
+    },
 
-            fillCellSelection: function(selection, complex, selectionExcludesMulti) {
-                var forms  = [],
-                    values = [],
-                    label;
+    _buildGemHtmlTooltip: function(lines, complex, context, gem, index) {
+      /*
+       * Multi-chart formulas are not shown in the tooltip
+       * They're on the small chart's title.
+       *
+       * Also, if the chart hides null members,
+       * don't show them in the tooltip.
+       * Using the scene's group, preferably, because the datum (here the complex) may have dimensions
+       * that are null in the groups' own atoms.
+       */
+      if(this._nonMultiGemFilter(gem) &&
+         !(this.chart._hideNullMembers && this._isNullMember(context.scene.group || complex, gem))) {
+        this.base.apply(this, arguments);
+      }
+    },
 
-                this.getSelectionGems(selectionExcludesMulti)
-                    .each(function(gem) {
-                        var atom = complex.atoms[gem.cccDimName];
-                        forms.push(gem.name);
+    fillCellSelection: function(selection, complex, selectionExcludesMulti) {
+      var forms = [],
+          values = [],
+          label;
 
-                        // Translate back null member values to the original member value,
-                        // which is accessible in rawValue.
-                        values.push(atom.value == null ? atom.rawValue : atom.value);
-                        label = atom.label; // TODO is this ok?
-                    });
+      this.getSelectionGems(selectionExcludesMulti)
+          .each(function(gem) {
+            var atom = complex.atoms[gem.cccDimName];
+            forms.push(gem.name);
 
-                if(forms.length) {
-                    var axisId = this.id;
-                    // Dummy property, just to force Analyzer to read the axis info
-                    selection[axisId] = true;
+            // Translate back null member values to the original member value,
+            // which is accessible in rawValue.
+            values.push(atom.value == null ? atom.rawValue : atom.value);
+            label = atom.label; // TODO is this ok?
+          });
 
-                    selection[axisId + 'Id'   ] = forms;
-                    selection[axisId + 'Item' ] = values;
-                    selection[axisId + 'Label'] = label;
-                }
-            },
+      if(forms.length) {
+        var axisId = this.id;
+        // Dummy property, just to force Analyzer to read the axis info
+        selection[axisId] = true;
 
-            getSelectionGems: function(selectionExcludesMulti) {
-                if(selectionExcludesMulti == null) selectionExcludesMulti = true;
+        selection[axisId + "Id"] = forms;
+        selection[axisId + "Item"] = values;
+        selection[axisId + "Label"] = label;
+      }
+    },
 
-                return def.query(this.gems)
-                     .where(function(gem) {
-                        return !gem.isMeasureDiscrim && (!selectionExcludesMulti || this._nonMultiGemFilter(gem));
-                     }, this);
-            }
-        }
-    });
+    getSelectionGems: function(selectionExcludesMulti) {
+      if(selectionExcludesMulti == null) selectionExcludesMulti = true;
+
+      return def.query(this.gems)
+          .where(function(gem) {
+            return !gem.isMeasureDiscrim && (!selectionExcludesMulti || this._nonMultiGemFilter(gem));
+          }, this);
+    }
+  });
 });
